(PC-17858) fix(accessibility): Make results page accessible
